requirements.txt
Pillow
text-blind-watermark
ffmpeg-python
fastapi
uvicorn[standard]
python-multipart
